requirements.txt
# MioTracker GUI Requirements
# Core GUI Framework
PyQt5>=5.15.0

# Data visualization and plotting
pyqtgraph>=0.12.0

# Scientific computing and data analysis
numpy>=1.21.0
pandas>=1.3.0

# Serial communication
pyserial>=3.5

# WebSocket communication
websocket-client>=1.3.0

# Optional: For better data export and analysis
matplotlib>=3.5.0
scipy>=1.7.0 